refactor: remove unused code and update exception handling in langflow.api.utils and langflow.exceptions.api
