+++ src/backend/base/langflow/api/v1/endpoints.py
@@ -261,7 +261,7 @@ async def simplified_run_flow(
         else:
             logger.exception(exc)
             raise APIException(
-                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, exception=from_exc_and_flow(exc, flow)
+                status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, exception=APIException.from_exc_and_flow(exc, flow)
             ) from exc
     except InvalidChatInputException as exc:
         logger.error(exc)
@@ -279,7 +279,7 @@ async def simplified_run_flow(
         )
         logger.exception(exc)
         raise APIException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, exception=from_exc_and_flow(exc, flow)
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, exception=APIException.from_exc_and_flow(exc, flow)
         ) from exc
